requirements.txt
autopep8==2.0.1
black==23.1.0
isort==5.11.5
mypy==1.0.0
numpy==1.21.6
pandas==1.3.5
psycopg2-binary==2.9.5
pytest==7.2.1
python-dotenv==0.21.1
requests==2.28.2
requests-mock==1.10.0
SQLAlchemy==2.0.3
tabulate==0.9.0
